+++ src/pages/companies/page/CompanyPage.tsx
@@ -5,24 +5,30 @@ import { Helmet } from "react-helmet";
 
 import { useScrollTopOnMount } from "src/shared/hooks/useScrollTopOnMount";
 import { useSearchQueryDef } from "src/shared/hooks/useSearchQueryDef";
+import { useSearchSuggestions } from "src/shared/hooks/useSearchSuggestions";
 import { useSearchSort } from "src/shared/hooks/useSearchSort";
+import { useSearchLocationFilter } from "src/shared/hooks/useSearchLocationFilter";
 import { useSearchSalaryFilter } from "src/shared/hooks/useSearchSalaryFilter";
 import { useSearchRatingFilter } from "src/shared/hooks/useSearchRatingFilter";
 import { useSearch } from "src/shared/hooks/useSearch";
 
-import { availableSortOptions, SearchType } from "src/shared/constants/search";
-
-import { GetJobDetails } from "../graphql/types/GetJobDetails";
-import { GetJobReviews } from "../graphql/types/GetJobReviews";
-import { GET_JOB_DETAILS, getJobReviewsQueryBuilder } from "../graphql/queries";
-import { buildJobDetails, buildJobReviewsCardList } from "../graphql/utils";
+import { GetCompanyDetails } from "../graphql/types/GetCompanyDetails";
+import { GetCompanyJobs } from "../graphql/types/GetCompanyJobs";
+import {
+  GET_COMPANY_DETAILS,
+  getCompanyJobsQueryBuilder,
+} from "../graphql/queries";
+import {
+  buildCompanyDetails,
+  buildCompanyJobCardsList,
+} from "../graphql/utils";
 
 import {
   PageContainer,
   SearchOptionsMenu,
   SearchResultCardDisplay,
 } from "src/components";
-import JobDetailsCard from "./JobDetailsCard";
+import CompanyDetailsCard from "../components/CompanyDetailsCard";
 
 /*******************************************************************
  *                  **Utility functions/constants**                *
@@ -30,91 +36,89 @@ import JobDetailsCard from "./JobDetailsCard";
 /**
  * Creates markup for the title in the tab bar.
  */
-const getTitleMarkup = (jobName?: string, companyName?: string) =>
-  jobName && companyName
-    ? `${jobName} at ${companyName} • intern+`
-    : "Job details • intern+";
+const getTitleMarkup = (companyName?: string) =>
+  companyName ? `${companyName} • intern+` : "Company details • intern+";
 
 /*******************************************************************
  *                           **Component**                         *
  *******************************************************************/
-const JobPage: React.FC = () => {
+const CompanyPage: React.FC = () => {
   useScrollTopOnMount();
 
-  const { jobId } = useParams();
+  const { companySlug } = useParams();
+  const searchSuggestions = useSearchSuggestions({ companySlug });
 
   /**
-   * Fetch the *details of the job* with corresponding id.
+   * Fetch *details of the company* with the corresponding slug.
    */
   const {
     loading: detailsLoading,
     error: detailsError,
     data: detailsData,
-  } = useQuery<GetJobDetails>(GET_JOB_DETAILS, {
-    variables: { id: jobId },
+  } = useQuery<GetCompanyDetails>(GET_COMPANY_DETAILS, {
+    variables: { slug: companySlug },
   });
 
-  const jobDetails = useMemo(
+  const companyDetails = useMemo(
     () =>
-      detailsData && detailsData.job
-        ? buildJobDetails(detailsData.job)
+      detailsData && detailsData.company
+        ? buildCompanyDetails(detailsData.company)
         : undefined,
     [detailsData]
   );
 
   /**
-   * Fetch *reviews of the job*.
+   * Fetch *jobs at the company*.
    */
-  const { QUERY_DEF } = useSearchQueryDef(getJobReviewsQueryBuilder);
+  const { QUERY_DEF } = useSearchQueryDef(getCompanyJobsQueryBuilder);
   const {
     // search info
     searchState,
     searchResults,
+    unfilteredResults,
 
     // callbacks
     triggerSearchNew,
     triggerSearchNextBatch,
-  } = useSearch<GetJobReviews>(
+  } = useSearch<GetCompanyJobs>(
     QUERY_DEF,
     {
       variables: {
-        id: jobId,
+        slug: companySlug,
       },
     },
-    buildJobReviewsCardList
+    buildCompanyJobCardsList
   );
 
   /**
    * For search options menu
    */
-  const sortOption = useSearchSort(availableSortOptions[SearchType.REVIEWS]);
+  const sortOption = useSearchSort();
   const salaryOption = useSearchSalaryFilter();
+  const locationOption = useSearchLocationFilter(unfilteredResults);
   const ratingOption = useSearchRatingFilter();
 
   return (
     <>
       <Helmet>
-        <title>
-          {getTitleMarkup(
-            jobDetails && jobDetails.name,
-            jobDetails && jobDetails.companyName
-          )}
-        </title>
+        <title>{getTitleMarkup(companyDetails && companyDetails.name)}</title>
       </Helmet>
 
-      <PageContainer id="job-page">
-        <JobDetailsCard
+      <PageContainer id="company-page">
+        <CompanyDetailsCard
           loading={detailsLoading}
           error={detailsError !== undefined}
-          jobDetails={jobDetails}
+          companyDetails={companyDetails}
           searchFieldProps={{
             onTriggerSearch: triggerSearchNew,
-            inputProps: { placeholder: "Find a review" },
+            suggestions: searchSuggestions,
+            inputProps: { placeholder: "Find a position" },
           }}
         />
 
         <SearchOptionsMenu
           sortOption={sortOption}
+          locationOption={locationOption}
           salaryOption={salaryOption}
           ratingOption={ratingOption}
           onOptionChange={() => triggerSearchNew(undefined, true)}
@@ -130,4 +134,4 @@ const JobPage: React.FC = () => {
   );
 };
 
-export default JobPage;
+export default React.memo(CompanyPage);
