+++ src/pages/error/404/index.tsx
@@ -22,6 +22,7 @@ const Contents = styled.div`
   & > div.textContainer {
     margin-left: 50px;
     max-width: 300px;
+    line-height: 1.3;
   }
 
   ${({ theme }) => theme.mediaQueries.tablet`
