+++ src/pages/reviews/page/index.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+import { withLoadingPage } from "src/shared/utils/withLoadingPage";
+
+const Review = React.lazy(() =>
+  import(/* webpackChunkName: "review" */ "./ReviewPage")
+);
+
+export default withLoadingPage(Review);
